service/owner-service/requirements.txt
fastapi[standard]
redis
python-dotenv
setuptools